# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "CRUD REST API Laravel 11",
        "description": "Documentation for Laravel 11 CRUD REST API",
        "contact": {
            "name": "API Support",
            "email": "admin@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://crud-rest-api-laravel-11.test/api",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/search/name": {
            "get": {
                "tags": [
                    "Search"
                ],
                "summary": "Search by name (Turner Mia)",
                "description": "Search for records with the name 'Turner Mia' in the external API",
                "operationId": "84e64cd68174fdf0f6ec6a5864f3d107",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "count": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "NAMA": {
                                                        "type": "string",
                                                        "example": "Turner Mia"
                                                    },
                                                    "NIM": {
                                                        "type": "string",
                                                        "example": "9352078461"
                                                    },
                                                    "YMD": {
                                                        "type": "string",
                                                        "example": "20230405"
                                                    },
                                                    "ALAMAT": {
                                                        "type": "string",
                                                        "example": "Apt. 398 36279 Langworth Trail, Lake Donatoville, TN 99242-0940"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Failed to fetch data from external API"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/search/nim": {
            "get": {
                "tags": [
                    "Search"
                ],
                "summary": "Search by NIM (9352078461)",
                "description": "Search for records with the NIM '9352078461' in the external API",
                "operationId": "24ef65c4716371acff8ca066b8824481",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "count": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "NAMA": {
                                                        "type": "string",
                                                        "example": "Turner Mia"
                                                    },
                                                    "NIM": {
                                                        "type": "string",
                                                        "example": "9352078461"
                                                    },
                                                    "YMD": {
                                                        "type": "string",
                                                        "example": "20230405"
                                                    },
                                                    "ALAMAT": {
                                                        "type": "string",
                                                        "example": "Apt. 398 36279 Langworth Trail, Lake Donatoville, TN 99242-0940"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Failed to fetch data from external API"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/search/ymd": {
            "get": {
                "tags": [
                    "Search"
                ],
                "summary": "Search by YMD (20230405)",
                "description": "Search for records with the YMD '20230405' in the external API",
                "operationId": "abcf3baf6d7ba4038616d125d380d17c",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "count": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "NAMA": {
                                                        "type": "string",
                                                        "example": "Turner Mia"
                                                    },
                                                    "NIM": {
                                                        "type": "string",
                                                        "example": "9352078461"
                                                    },
                                                    "YMD": {
                                                        "type": "string",
                                                        "example": "20230405"
                                                    },
                                                    "ALAMAT": {
                                                        "type": "string",
                                                        "example": "Apt. 398 36279 Langworth Trail, Lake Donatoville, TN 99242-0940"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Failed to fetch data from external API"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}